# openapi.yaml
...
openapi: 3.0.3
info:
  title: ScholarAI API
  description: AI-powered scholarship matching platform
  version: 5.0.0
  contact:
    name: ScholarAI Team
    email: support@scholarai.com
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://api.scholarai.com
    description: Production server
paths:
...
  /scholarship/scrape:
    post:
      summary: Scrape and normalize a single scholarship from URL
      tags:
        - Scholarship Scraping
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  format: uri
                  description: URL of the scholarship page to scrape
              required:
                - url
      responses:
        '200':
          description: Scholarship scraped and normalized successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Scholarship'
                  metadata:
                    $ref: '#/components/schemas/ScrapingMetadata'
                  timestamp:
                    type: string
                    format: date-time
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Failed to scrape or normalize
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /scholarship/batch-scrape:
    post:
      summary: Scrape and normalize multiple scholarships from URLs
      tags:
        - Scholarship Scraping
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                urls:
                  type: array
                  items:
                    type: string
                    format: uri
                  description: Array of scholarship page URLs to scrape
                  maxItems: 50
                concurrency:
                  type: integer
                  minimum: 1
                  maximum: 10
                  default: 3
                  description: Number of concurrent scraping requests
              required:
                - urls
      responses:
        '200':
          description: Batch scraping completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  scholarships:
                    type: array
                    items:
                      $ref: '#/components/schemas/Scholarship'
                  summary:
                    $ref: '#/components/schemas/BatchSummary'
                  errors:
                    type: array
                    items:
                      $ref: '#/components/schemas/ScrapingError'
                  timestamp:
                    type: string
                    format: date-time
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /scholarship/validate:
    post:
      summary: Validate a scholarship object against the schema
      tags:
        - Scholarship Scraping
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                scholarship:
                  $ref: '#/components/schemas/Scholarship'
              required:
                - scholarship
      responses:
        '200':
          description: Validation completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  valid:
                    type: boolean
                  errors:
                    type: array
                    items:
                      type: string
                  scholarship:
                    $ref: '#/components/schemas/Scholarship'
                  timestamp:
                    type: string
                    format: date-time
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /scholarship/test:
    get:
      summary: Test the scholarship scraping functionality
      tags:
        - Scholarship Scraping
      responses:
        '200':
          description: Test completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  testResults:
                    type: object
                  timestamp:
                    type: string
                    format: date-time
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
        details:
          type: array
          items:
            type: object
          description: Additional error details (e.g., validation errors)
...
    Scholarship:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the scholarship
        name:
          type: string
          description: Name of the scholarship
        country:
          type: string
          nullable: true
          description: Country where the scholarship is available
        degree:
          type: string
          enum: ["Bachelor", "Master", "PhD", "Any"]
          description: Degree level for the scholarship
        eligibility:
          type: string
          maxLength: 220
          description: Eligibility requirements (max 220 characters)
        deadline:
          type: string
          nullable: true
          pattern: '^(\\d{4}-\\d{2}-\\d{2}|varies)$'
          description: Application deadline in YYYY-MM-DD format or "varies"
        link:
          type: string
          format: uri
          description: Official application or information URL
        source:
          type: string
          description: Source domain of the scholarship information
        isFullyFunded:
          type: boolean
          description: Whether the scholarship is fully funded
      required:
        - id
        - name
        - degree
        - eligibility
        - link
        - source
        - isFullyFunded
...
    ScrapingMetadata:
      type: object
      properties:
        url:
          type: string
          format: uri
          description: Original URL scraped
        scrapedAt:
          type: string
          format: date-time
          description: Timestamp when scraping occurred
        processingTime:
          type: number
          description: Time taken to scrape and process in seconds
        source:
          type: string
          description: Domain or source identifier
      required:
        - url
        - scrapedAt
        - processingTime
        - source
    BatchSummary:
      type: object
      properties:
        total:
          type: integer
          description: Total number of URLs processed
        successful:
          type: integer
          description: Number of successfully scraped scholarships
        failed:
          type: integer
          description: Number of failed scraping attempts
        processingTime:
          type: number
          description: Total processing time in seconds
      required:
        - total
        - successful
        - failed
        - processingTime
    ScrapingError:
      type: object
      properties:
        url:
          type: string
          format: uri
          description: URL that failed to scrape
        error:
          type: string
          description: Error message
        errorType:
          type: string
          enum: ["network", "parsing", "validation", "timeout"]
          description: Type of error encountered
        timestamp:
          type: string
          format: date-time
          description: When the error occurred
      required:
        - url
        - error
        - errorType
        - timestamp
...